requirements.txt
python-dotenv~=1.1.0
langchain~=0.3.25
openai~=1.86.0
requests~=2.32.3
streamlit